fix: убрать подчеркивание инпута в форме добавления карточки, перенести код в index.js, создать функцию createCard
