+++ blocks/popup/__input/_type_error/popup__input_type_error.css
@@ -1,0 +1,3 @@
+.popup__input_type_error {
+  border-bottom-color: #ff0000;
+}
